requirements.txt
requests==2.31.0
python-dotenv==1.0.0
selenium==4.16.0 
onetimepass==1.0.1
flask==3.0.3
